Abolish Event class hierarchy (No Event subclasses) - events.py turned into events/_events.py and events/resolve.py
