+++ pooltool/events/__init__.py
@@ -1,0 +1,3 @@
+#! /usr/bin/env python
+
+from pooltool.events._events import Event, Events, EventType, NonEvent, event_resolvers
